refactor: remove duplicated component DisplayField
