--- src/components/ToolMenu/FeatureInfo/DisplayField/index.spec.tsx
@@ -1,9 +1,0 @@
-import DisplayField from '.';
-
-describe('<DisplayField />', () => {
-
-  it('is defined', () => {
-    expect(DisplayField).toBeDefined();
-  });
-
-});
