+++ src/components/ToolMenu/FeatureInfo/FeatureInfoForm/index.tsx
@@ -17,7 +17,7 @@ import {
   PropertyFormItemReadConfig
 } from '@terrestris/shogun-util/dist/model/Layer';
 
-import DisplayField from '../DisplayField';
+import DisplayField from '../../../DisplayField';
 
 export type FeatureInfoFormProps = FormProps & {
   formConfig?: PropertyFormItemReadConfig[];
